requirements.txt
pyzmq  # inter process communication library
pyyaml  # configuration parsing
psutil  # to get process stats
pyqtgraph>=0.13  # plotting library
pyqt5==5.15.2  # https://stackoverflow.com/questions/65447314/attributeerror-module-sipbuild-api-has-no-attribute-prepare-metadata-for-bui/65449548#65449548
pytest
testfixtures  # for unittests with logging output inspection
# coverage  # for unit test coverage stats, only needed for CI autocoverage
# coveralls  # for unit test coverage stats, only needed for CI autocoverage
